+++ src/gcloud_functions/requirements.txt
@@ -1,0 +1,3 @@
+Pillow==9.0.0
+torch==1.10.1
+torchvision==0.11.2
